+++ lib/lunar.ml
@@ -1,0 +1,17 @@
+(* Copyright (c) 2026, Cargocut and the Lunar developers.
+   All rights reserved.
+
+   SPDX-License-Identifier: BSD-3-Clause *)
+
+module Era = Era
+module Month = Month
+module Weekday = Weekday
+module Time = Time
+module Date = Date
+module Datetime = Datetime
+module Timezone = Timezone
+module Zoned_datetime = Zoned_datetime
+module Resolution = Resolution
+module Duration = Duration
+module Range = Range
+module Sigs = Sigs